requirements.txt
tensorflow
streamlit
pillow
numpy
torch